// src/users/dto/create-user.dto.ts
import { ApiProperty } from "@nestjs/swagger";

export class CreateUserDto {

  readonly id?: number;

  @ApiProperty({example: '22@fff.com', description: 'valid email address'})
  readonly email: string;

  @ApiProperty({example: 'dKsr&jdl3', description: 'valid password'})
  readonly password: string;

  @ApiProperty({example: 1, description: 'id referrence user'})
  readonly refToUser: number;

  @ApiProperty({example: 1, description: 'id referrence user'})
  readonly role?: 'Guest' | 'Investor' | 'Owner';

}